+++ jest.config.ts
@@ -20,8 +20,8 @@ const config: Config.InitialOptions = {
     '^constants/(.*)$': '<rootDir>/src/constants/$1',
     '^components$': '<rootDir>/src/components',
     '^content$': '<rootDir>/src/content',  
-    '^context$': '<rootDir>/src/contexts',
-    '^contexts/(.*)$': '<rootDir>/src/contexts/$1',
+    '^context$': '<rootDir>/src/context',
+    '^context/(.*)$': '<rootDir>/src/context/$1',
   },
   
   setupFilesAfterEnv: ['<rootDir>/src/tests/setupTests.ts'],
