+++ src/components/Logo/Logo.test.tsx
@@ -4,7 +4,7 @@ import { Logo } from "./Logo";
 import { ThemeDark } from "themes";
 import { ThemeProvider } from "styled-components";
 import { ReactNode } from "react";
-import ScrollProvider from "contexts/ScrollContext";
+import ScrollProvider from "context/ScrollContext";
 
 const renderWithTheme = (component: ReactNode) => {
  return render(
