+++ src/components/Timeline/Timeline.test.tsx
@@ -3,7 +3,7 @@ import { Timeline } from "./Timeline";
 import { ThemeDark } from "themes";
 import { ThemeProvider } from "styled-components";
 import { ReactNode } from "react";
-import ScrollProvider from "contexts/ScrollContext";
+import ScrollProvider from "context/ScrollContext";
 
 const renderWithTheme = (component: ReactNode) => {
  return render(
