--- src/contexts/index.ts
@@ -1,1 +1,0 @@
-export { default as ScrollContextProvider } from "./ScrollContext";
